pyproject.toml
[project]
name = "coglet"
version = "0.1.0"
description = "Fractal asynchronous control for distributed agent systems"
requires-python = ">=3.12,<3.13"
dependencies = [
    "mettagrid",
    "pydantic",
    "fastapi",
    "uvicorn",
    "fastapi-mcp",
    "click>=8.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
]

[project.scripts]
coglet = "coglet.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/coglet", "cogs/cogames"]

[tool.ty.environment]
extra-paths = ["src", "."]
